+++ NEWS.md
@@ -1,3 +1,10 @@
+# gamvinereg 0.2.0 (June 20, 2024)
+
+-   Added function `cpdf` for conditional density function.
+-   Adapted function call from `pit` to `cpit`.
+-   Slightly improved code. 
+
+
 # gamvinereg 0.1.0 (September 12, 2023)
 
 -   First release.